got the reorganization working without errors for one component. Now we have to reorganize every component from the app
