+++ src/front/js/layout.js
@@ -6,15 +6,16 @@ import { BackendURL } from "./component/backendURL";
 import { Home } from "./pages/home";
 import { ContactUs } from "./pages/contactUs";
 import { Single } from "./pages/single";
-import { ProviderProfile } from "./pages/providerProfile";
-import { ProviderMapPage } from "./pages/providerMapPage";
 import injectContext from "./store/appContext";
 import { ProviderSignUp } from "./pages/providerSignUp";
+import { ProviderProfile } from "./pages/providerProfile";
+import { ProviderMapPage } from "./pages/providerMapPage";
 import { Navbar } from "./component/navbar";
 import { Footer } from "./component/footer";
+
 import { ProviderServices } from "./pages/providerServices";
 import { ProviderProfileCompleted } from "./pages/providerProfileCompleted";
-import { ClientMapPage } from "./pages/clientMapPage";
+import { ClientMapPage } from "./pages/clientAccountCreationFlow/clientMapPage";
 import { ClientSignUp1 } from "./pages/clientSignUp1";
 import { ClientSignUp2 } from "./pages/clientSignUp2";
 import { ClientServices1 } from "./pages/clientServices1";
