+++ src/front/js/pages/clientAccountCreationFlow/clientMapPage.js
@@ -1,6 +1,6 @@
 import React, { useContext, useState } from "react";
-import { Context } from "../store/appContext";
-import "../../styles/home.css";
+import { Context } from "../../store/appContext";
+import "../../../styles/home.css";
 import { useNavigate } from "react-router-dom";
 
 export const ClientMapPage = () => {
